Remove the dms.Client wrapper; call the SDK directly
Delete the Client struct and client.go. DeploymentState and OperationBuffer
now hold bundledeployments.BundleDeploymentsInterface, and each SDK call is
made at its call site. libs/dms keeps only the non-trivial builders that would
otherwise push wire details into caller packages: the ForceSendFields
update-mask builders, ListResources, and the name/prefix helpers. Pure
refactor - acceptance goldens are unchanged.
