+++ bundle/direct/dstate/state.go
@@ -122,9 +122,9 @@ type DeploymentState struct {
 	// history, in which case InitializeOperationBuffer installs it once the version exists.
 	operationBuffer *dms.OperationBuffer
 
-	// dmsClient talks to the deployment metadata service. Open builds it from the workspace
+	// dmsService talks to the deployment metadata service. Open sets it from the workspace
 	// client when the deployment records history; nil otherwise.
-	dmsClient *dms.Client
+	dmsService bundledeployments.BundleDeploymentsInterface
 
 	// versionCompleted makes CompleteVersion a no-op after the first call, so a deferred safety-net
 	// completion after an explicit one does nothing.
@@ -204,7 +204,7 @@ type WALEntry struct {
 // (after approval) - which is why it is not an Open option. It also records the id a first deploy
 // just created, which Open could not know, so CompleteVersion later has it.
 func (db *DeploymentState) InitializeOperationBuffer(ctx context.Context, deploymentID string, versionID int) {
-	buf := dms.StartOperationBuffer(ctx, db.dmsClient, deploymentID, versionID)
+	buf := dms.StartOperationBuffer(ctx, db.dmsService, deploymentID, versionID)
 
 	db.mu.Lock()
 	defer db.mu.Unlock()
@@ -248,7 +248,7 @@ func (db *DeploymentState) CompleteVersion(ctx context.Context, success bool) (b
 		return false, nil
 	}
 	db.versionCompleted = true
-	deploymentID, client, versionID := db.DeploymentID, db.dmsClient, db.VersionID
+	deploymentID, service, versionID := db.DeploymentID, db.dmsService, db.VersionID
 	db.mu.Unlock()
 
 	// A recording failure fails the version even when the caller counted the deploy a success: the
@@ -262,7 +262,11 @@ func (db *DeploymentState) CompleteVersion(ctx context.Context, success bool) (b
 	if !success {
 		reason = bundledeployments.VersionCompleteVersionCompleteFailure
 	}
-	if err := client.CompleteVersion(ctx, deploymentID, versionID, reason); err != nil {
+	_, err := service.CompleteVersion(ctx, bundledeployments.CompleteVersionRequest{
+		Name:             dms.VersionName(deploymentID, versionID),
+		CompletionReason: reason,
+	})
+	if err != nil {
 		return false, err
 	}
 	log.Infof(ctx, "Completed deployment version: deployment=%s version=%d reason=%s", deploymentID, versionID, reason)
@@ -450,10 +454,10 @@ func (db *DeploymentState) IsDeploymentMetadataService() bool {
 	return db.isDeploymentMetadataService()
 }
 
-// DmsClient returns the deployment metadata service client Open built from the workspace client,
+// DmsService returns the deployment metadata service client Open set from the workspace client,
 // or nil when the deployment does not record history.
-func (db *DeploymentState) DmsClient() *dms.Client {
-	return db.dmsClient
+func (db *DeploymentState) DmsService() bundledeployments.BundleDeploymentsInterface {
+	return db.dmsService
 }
 
 // GetOrInitLineage returns the deployment lineage, generating and storing a new
@@ -617,15 +621,11 @@ To record this bundle's history, start it over as a new deployment:
 	if recorded {
 		// The service is the source of truth for a recorded deployment; the file is a tombstone
 		// carrying only the marker, and applyDMSState loads the resources the service holds.
-		client, err := dms.NewClient(cmdctx.WorkspaceClient(ctx))
-		if err != nil {
-			return err
-		}
-		db.dmsClient = client
+		db.dmsService = cmdctx.WorkspaceClient(ctx).BundleDeployments
 		db.DeploymentID = dmsDeployment.DeploymentID
 
 		if dmsDeployment.DeploymentID != "" {
-			resources, err := db.dmsClient.ListResources(ctx, dmsDeployment.DeploymentID)
+			resources, err := dms.ListResources(ctx, db.dmsService, dmsDeployment.DeploymentID)
 			if err != nil {
 				return err
 			}
