+++ bundle/phases/dms.go
@@ -16,15 +16,16 @@ import (
 	"github.com/databricks/cli/libs/log"
 	"github.com/databricks/cli/libs/logdiag"
 	"github.com/databricks/cli/libs/workspaceurls"
+	"github.com/databricks/databricks-sdk-go/common/types/fieldmask"
 	"github.com/databricks/databricks-sdk-go/service/bundledeployments"
 )
 
 // stagedOperations lists the resources the plan will touch, for CreateVersion to stage an
 // operation each. Skipped and undefined actions are left out: nothing is applied for them, so
 // their operations would stay pending and the service would hold no state for them.
-func stagedOperations(plan *deployplan.Plan) ([]dms.StagedOperation, error) {
+func stagedOperations(plan *deployplan.Plan) ([]bundledeployments.StagedOperation, error) {
 	actions := plan.GetActions()
-	staged := make([]dms.StagedOperation, 0, len(actions))
+	staged := make([]bundledeployments.StagedOperation, 0, len(actions))
 	for _, action := range actions {
 		if action.ActionType == deployplan.Skip || action.ActionType == deployplan.Undefined {
 			continue
@@ -33,8 +34,8 @@ func stagedOperations(plan *deployplan.Plan) ([]dms.StagedOperation, error) {
 		if err != nil {
 			return nil, fmt.Errorf("%s: %w", action.ResourceKey, err)
 		}
-		staged = append(staged, dms.StagedOperation{
-			ResourceKey: action.ResourceKey,
+		staged = append(staged, bundledeployments.StagedOperation{
+			ResourceKey: strings.TrimPrefix(action.ResourceKey, dms.StatePrefix),
 			ActionType:  actionType,
 		})
 	}
@@ -69,19 +70,31 @@ func actionToSDK(a deployplan.ActionType) (bundledeployments.OperationActionType
 // no deployment behind; a first deploy's new id is then stamped into the plan (StampDeploymentID).
 func createOrUpdateDeployment(ctx context.Context, b *bundle.Bundle, current *bundledeployments.Deployment) {
 	db := &b.DeploymentBundle
-	dmsClient := db.StateDB.DmsClient()
+	dmsService := db.StateDB.DmsService()
 	metadata := deploymentMetadata(b)
 	deploymentID := db.StateDB.DeploymentID
 	if deploymentID == "" {
-		id, err := dmsClient.CreateDeployment(ctx, b.Config.Workspace.StatePath, metadata)
+		dep := metadata.Deployment()
+		dep.InitialParentPath = b.Config.Workspace.StatePath
+		created, err := dmsService.CreateDeployment(ctx, bundledeployments.CreateDeploymentRequest{Deployment: dep})
 		if err != nil {
 			logdiag.LogError(ctx, fmt.Errorf("failed to create deployment: %w", err))
 			return
 		}
-		deploymentID = id
+		// The server assigns the id as the workspace node it creates under the parent path.
+		deploymentID, err = dms.DeploymentIDFromName(created.Name)
+		if err != nil {
+			logdiag.LogError(ctx, fmt.Errorf("failed to create deployment: %w", err))
+			return
+		}
 		db.StateDB.DeploymentID = deploymentID
 	} else if mask := metadata.StaleFields(current); mask != "" {
-		if err := dmsClient.UpdateDeployment(ctx, deploymentID, metadata, mask); err != nil {
+		_, err := dmsService.UpdateDeployment(ctx, bundledeployments.UpdateDeploymentRequest{
+			Name:       dms.DeploymentName(deploymentID),
+			Deployment: dms.DeploymentUpdate(metadata, mask),
+			UpdateMask: fieldmask.FieldMask{Paths: strings.Split(mask, ",")},
+		})
+		if err != nil {
 			logdiag.LogError(ctx, fmt.Errorf("failed to update deployment: %w", err))
 			return
 		}
@@ -96,10 +109,10 @@ func createOrUpdateDeployment(ctx context.Context, b *bundle.Bundle, current *bu
 // startVersion claims the version the run settled on and opens the buffer that records
 // each state write under it. Called after approval, so a declined deploy never claims a number.
 // A no-op when the bundle does not record deployment history.
-func startVersion(ctx context.Context, b *bundle.Bundle, versionType dms.VersionType, staged []dms.StagedOperation) error {
+func startVersion(ctx context.Context, b *bundle.Bundle, versionType dms.VersionType, staged []bundledeployments.StagedOperation) error {
 	db := &b.DeploymentBundle
-	dmsClient := db.StateDB.DmsClient()
-	if dmsClient == nil {
+	dmsService := db.StateDB.DmsService()
+	if dmsService == nil {
 		return nil
 	}
 	deploymentID := db.StateDB.DeploymentID
@@ -124,12 +137,16 @@ func startVersion(ctx context.Context, b *bundle.Bundle, versionType dms.Version
 			OriginUrl: git.OriginURL,
 		}
 	}
-	version, err := dmsClient.CreateVersion(ctx, deploymentID, versionID, dms.CreateVersionRequest{
-		CliVersion:        build.GetInfo().Version,
-		VersionType:       versionType,
-		PreviousVersionId: previousVersionID,
-		Operations:        staged,
-		GitInfo:           gitInfo,
+	version, err := dmsService.CreateVersion(ctx, bundledeployments.CreateVersionRequest{
+		Parent:    dms.DeploymentName(deploymentID),
+		VersionId: strconv.Itoa(versionID),
+		Version: bundledeployments.Version{
+			CliVersion:        build.GetInfo().Version,
+			VersionType:       versionType,
+			PreviousVersionId: previousVersionID,
+			GitInfo:           gitInfo,
+			Operations:        staged,
+		},
 	})
 	if err != nil {
 		return fmt.Errorf("failed to create deployment version: %w", err)
