+++ libs/dms/operation.go
@@ -3,11 +3,17 @@ package dms
 import (
 	"encoding/json"
 	"fmt"
+	"strconv"
 	"unicode/utf8"
 
 	"github.com/databricks/cli/libs/diag"
 	"github.com/databricks/databricks-sdk-go/service/bundledeployments"
 )
+
+// statePrefix is what a bundle state key carries and a DMS resource key does not: state calls a
+// job "resources.jobs.foo", DMS calls it "jobs.foo". Exported names take the state form; the
+// prefix comes off where a request is built, and back on where a resource is read.
+const StatePrefix = "resources."
 
 // maxStateSize is the largest serialized state DMS accepts per operation. More than this
 // and the resource cannot be recorded at all, so the deploy fails rather than leaving the
@@ -139,4 +145,38 @@ func (u OperationUpdate) Merge(newer OperationUpdate) OperationUpdate {
 	}
 
 	return merged
+}
+
+// newOperationUpdate builds the operation carrying exactly the fields update.Fields masks, plus
+// the sequence_id precondition. The service requires a masked field to be present and reads an
+// empty value as a write (error_message="" clears it), so masked fields that can be empty are
+// forced onto the wire; sequence_id is always sent and a freshly staged operation sits at 0,
+// which omitempty would drop. State is the exception: an absent value is how the service is told
+// the resource is gone, so a nil state is left off (never forced) while still named in the mask.
+func newOperationUpdate(update OperationUpdate, sequenceID string) (bundledeployments.Operation, error) {
+	sequence, err := strconv.ParseInt(sequenceID, 10, 64)
+	if err != nil {
+		return bundledeployments.Operation{}, fmt.Errorf("invalid sequence id %q: %w", sequenceID, err)
+	}
+
+	operation := bundledeployments.Operation{
+		SequenceId:      sequence,
+		ForceSendFields: []string{"SequenceId"},
+	}
+	if update.Fields.Has(FieldState) && update.State != nil {
+		operation.State = string(update.State)
+	}
+	if update.Fields.Has(FieldErrorMessage) {
+		operation.ErrorMessage = update.ErrorMessage
+		operation.ForceSendFields = append(operation.ForceSendFields, "ErrorMessage")
+	}
+	if update.Fields.Has(FieldResourceID) {
+		operation.ResourceId = update.ResourceID
+		operation.ForceSendFields = append(operation.ForceSendFields, "ResourceId")
+	}
+	if update.Fields.Has(FieldStatus) {
+		operation.Status = update.Status
+		operation.ForceSendFields = append(operation.ForceSendFields, "Status")
+	}
+	return operation, nil
 }
